fix(aarch64): 修复浮点开方的 Apple 严格构建
在舍入指数计算中显式转换额外精度位，消除 iOS 与 watchOS 开启 -Wsign-conversion 后的编译错误。

涉及文件：guest/aarch64/scalar-fp.h；不改变指令语义，仅统一整数运算类型。

+++ guest/aarch64/scalar-fp.h
@@ -1022,7 +1022,7 @@ static inline struct aarch64_scalar_fp_result aarch64_scalar_fp_sqrt(
     int result_exponent = (exponent - (exponent & 1)) / 2;
     struct aarch64_scalar_fp_result result = aarch64_scalar_fp_round(
             false, root,
-            result_exponent - (int) format.fraction_bits - extra_bits,
+            result_exponent - (int) format.fraction_bits - (int) extra_bits,
             remainder != 0, &format, fpcr);
     result.exceptions |= exceptions;
     return result;
